# Conflicts:
#	src/main/java/com/nutriplus/NutriPlusBack/repositories/ApplicationFoodRepository.java
#	src/main/resources/patients.graphql

+++ src/main/java/com/nutriplus/NutriPlusBack/domain/food/FoodModel.java
@@ -14,7 +14,7 @@ public abstract class FoodModel extends AbstractEntity {
     String foodGroup;
     double measureTotalGrams;  // in grams
     String measureType;        // homemade measure, such as a "tablespoon" or a "cup of tea"
-    int measureAmount;         // amount of measureType to reach measureTotalGrams
+    Double measureAmount;         // amount of measureType to reach measureTotalGrams
     boolean custom;            // is this food a personal customization of another food?
     boolean created;           // is this food a personal food created by some nutritionist?
     @Relationship(type = "CUSTOMIZE", direction = Relationship.OUTGOING)
